Completed the body of the GameCotroller class and more
Added methods to most of the classes and modified others, changed GameBoard name to DeprecatedGameBoard and created a new Borad class in its place. Moved method to pause until enter from ConsoleInOut to Pause class. From now on GameController should guide the flow of the app and comunicate model classes with ConsoleView, where all the methods are to be implemented yet